fix: update unit tests to work with Python 3.10+
The Python interpeter normalized quotes across all outputs, which
made some of the unit tests that look at interpreter output quietly
fail.  Update the strings under test to use the new style, and add
skip annotations for older Pythons, since at this point they are
known to be tested, and on the path to EOL.

NOTE: The `doctest` in the Python stdlib from 3.10+ has introduced
a bug in `doctest.OutputChecker.output_difference` where it stringifies
the difference into a single line.  Nobody seems to have noticed,
nor cared about it, but the tests here need to account for it.

Fixes: Issue #2